Фильтры нескольких компонентов, оптимизация

+++ CustomEcs/Entity.cs
@@ -1,16 +1,18 @@
 ﻿using System;
+using System.Diagnostics.Eventing.Reader;
 
 namespace CustomEcs
 {
     public class Entity
     {
-        static int defaultSizeBuffer = 8;
+        static readonly int defaultSizeBuffer = 8;
 
         readonly ComponentContainer container;
+        readonly MainClassECS mainClass;
         int[] IndexesComponents;
         public bool IsAlive { get; private set; }
         public int IndexEntity { get; private set; }
-        public Entity(int indexEntity)
+        internal Entity(int indexEntity, MainClassECS mainClass)
         {
             IndexesComponents = new int[defaultSizeBuffer * 2];
             for (int i = 0; i < IndexesComponents.Length; i++)
@@ -20,11 +22,12 @@ namespace CustomEcs
             IsAlive = true;
             container = ComponentContainer.GetInstance();
             this.IndexEntity = indexEntity;
+            this.mainClass = mainClass;
         }
 
         internal Entity ActivateEntity(int indexEntity)
         {
-            IndexesComponents = new int[defaultSizeBuffer * 2];
+            //IndexesComponents = new int[defaultSizeBuffer * 2];
             for (int i = 0; i < IndexesComponents.Length; i++)
             {
                 IndexesComponents[i] = -1;
@@ -37,6 +40,17 @@ namespace CustomEcs
         public void DeleteEntity()
         {
             IsAlive = false;
+            //Перебираем индексы типов структур и удаляем компоненты
+            for (int i = 0; i < IndexesComponents.Length; i += 2)
+            {
+                if(IndexesComponents[i] >= 0)
+                {
+                    int indexComponents = i + 1;
+                    container.GetComponent(IndexesComponents[i]).DeleteComponent(indexComponents);
+                }
+                
+            }
+            mainClass.lastDeletedEntity = IndexEntity;
         }
 
         public ref T AddOrGetComponent<T>() where T : struct
@@ -44,18 +58,23 @@ namespace CustomEcs
             Component<T> componentClass = Component<T>.GetInstanceComponent();
             int HashType = componentClass.HashType;
             int indexNewComponents = -1;
+
+            //Перебираем индексы типов структур
             for (int i = 0; i < IndexesComponents.Length; i += 2)
             {
+                //Если структура уже прикреплена то возвращем ссылку на нее
                 if (IndexesComponents[i] == HashType)
                 {
                     int indexComponents = i + 1;
                     return ref componentClass.GetComponent(IndexesComponents[indexComponents], IndexEntity);
                 }
-                if (IndexesComponents[i] == -1)
+                //Находим первый попавшийся свободный индекс
+                if (IndexesComponents[i] == -1 && indexNewComponents == -1)
                 {
                     indexNewComponents = i;
                 }
             }
+            //Если компонент не найден то создаем компонент на месте найденного индекса иначе расширяем контейнер
             if (indexNewComponents >= 0)
             {
                 int indexComponents = indexNewComponents + 1;
@@ -72,10 +91,12 @@ namespace CustomEcs
             }
         }
 
+        //Удаление компонента по типу структуры
         public void DeleteComponent<T>() where T : struct
         {
             Component<T> componentClass = Component<T>.GetInstanceComponent();
             int HashType = componentClass.HashType;
+            bool componentsExist = false;
             for (int i = 0; i < IndexesComponents.Length; i += 2)
             {
                 if (IndexesComponents[i] == HashType)
@@ -84,19 +105,16 @@ namespace CustomEcs
                     int indexComponents = i + 1;
                     componentClass.DeleteComponent(in IndexesComponents[indexComponents]);
                 }
-            }
-
-            for (int i = 0; i < IndexesComponents.Length; i += 2)
-            {
-                if (IndexesComponents[i] >= 0)
+                else if (IndexesComponents[i] >= 0)
                 {
-                    return;
+                    componentsExist = true;
                 }
             }
-
+            if (componentsExist) { return; }
             DeleteEntity();
         }
 
+        //Проверка прикреплены ли к сущности компоненты заданного типа
         internal bool CheckComponentType(int HashType)
         {
             for (int i = 0; i < IndexesComponents.Length; i += 2)
